First test of exercise 10 code changes
