+++ pyproject.toml
@@ -11,6 +11,7 @@ python-dotenv = "^0.14.0"
 requests = "^2.28.1"
 gunicorn = "^20.1.0"
 pytest = "^7.2.1"
+pymongo = "^4.3.3"
 
 [tool.poetry.dev-dependencies]
