+++ todo_app/app.py
@@ -1,5 +1,6 @@
 from flask import Flask, redirect,request,render_template
-from todo_app.data.trello_items import get_items, add_item, update_item
+#from todo_app.data.trello_items import get_items, add_item, update_item
+from todo_app.data.todo_items import get_items, add_item, update_item
 from todo_app.flask_config import Config
 def create_app():
     app = Flask(__name__)
