+++ todo_app/templates/index.html
@@ -3,7 +3,7 @@
 {% block content %}
   <div class="jumbotron">
     <h1 class="display-4"><b>To-Do App</b></h1>
-    <p class="lead"><b>Trello to-do app manager</b></p>  
+    <p class="lead"><b>To-do app manager</b></p>  
   </div>
   <div class="row">
     <div class="col-md-12">
